# Conflicts:
#	src/main/java/com/calendar/project/controller/EventController.java
#	src/main/java/com/calendar/project/dao/impl/EventDaoImpl.java
#	src/main/java/com/calendar/project/service/EventService.java
#	src/main/java/com/calendar/project/service/RoleService.java
#	src/main/java/com/calendar/project/service/impl/EventServiceImpl.java
#	src/main/webapp/static/showEvent.jsp
#	src/main/webapp/static/userControlPanel.jsp

+++ src/main/java/com/calendar/project/controller/EventController.java
@@ -1,5 +1,6 @@
 package com.calendar.project.controller;
 
+import org.springframework.http.MediaType;
 import com.calendar.project.dao.UserDao;
 import com.calendar.project.service.SecurityService;
 import com.calendar.project.service.UserService;
@@ -11,9 +12,13 @@ import com.calendar.project.model.User;
 import com.calendar.project.service.EventService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
-
 import java.util.*;
 import java.util.stream.Collectors;
 
@@ -32,7 +37,6 @@ public class EventController {
     @Autowired
     private UserService userService;
 
-
     @RequestMapping(value = "/events", method = RequestMethod.GET)
     public String showAllEvents(Model model) {
         model.addAttribute("events", eventService.getAllEvents());
@@ -41,7 +45,7 @@ public class EventController {
     }
 
     @RequestMapping(value = "/updateEvent", method = RequestMethod.GET)
-    public String updateEvent(Long eventId, Model model) {
+    public String updateEvent(int eventId, Model model){
         model.addAttribute("eventForm", eventService.getEvent(eventId));
 
         return "updateEvent";
@@ -63,7 +67,7 @@ public class EventController {
     }
 
     @RequestMapping(value = "/deleteEvent", method = RequestMethod.GET)
-    public String deleteEvent(Long eventId, Model model) {
+    public String deleteEvent(int eventId, Model model){
         model.addAttribute("eventForm", eventService.getEvent(eventId));
 
         return "deleteEvent";
@@ -75,6 +79,7 @@ public class EventController {
 
         return "redirect:/userPage";
     }
+
 
     @RequestMapping(value = "/createEvent", method = RequestMethod.GET)
     public String createEvent(Model model) {
@@ -104,8 +109,9 @@ public class EventController {
     }
 
     @RequestMapping(value = "/showEvent", method = RequestMethod.GET)
-    public String showEvent(Model model, Long eventId) {
+    public String showEvent(Model model, int eventId){
         Event event = eventService.getEvent(eventId);
+
         model.addAttribute("eventForm", event);
 
         return "showEvent";
@@ -120,4 +126,3 @@ public class EventController {
         return participantsByEvent;
     }
 }
-
